requirements.txt
playwright>=1.40.0
python-pptx>=0.6.21
pyyaml>=6.0
